Moved App, removed debounce
