+++ autoscheduler/frontend/src/components/App/App.css.d.ts
