+++ autoscheduler/frontend/src/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as ReactDom from 'react-dom';
 
-import App from './App/App';
+import App from './components/App/App';
 
 ReactDom.render(
   <App />,
