+++ autoscheduler/frontend/src/tests/App.test.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import * as React from 'react';
 
-import App from '../App/App';
+import App from '../components/App/App';
 
 test('renders without errors', () => {
   const { container } = render(<App />);
